add Flow piece constraint + change notation int to floatwq

+++ src/abstracts/AbstractChordConstraint.h
@@ -11,7 +11,7 @@ class AbstractChordConstraint
 	public:
 		int getCoeff();
 		void setCoeff(int _coeff);
-		virtual int eval(Chord* _chord) = 0;
+		virtual float eval(Chord* _chord) = 0;
 };
 
 #endif //ABSTRACT_CHORD_CONSTRAINT_H
