+++ src/abstracts/AbstractPieceConstraint.h
@@ -11,7 +11,7 @@ class AbstractPieceConstraint
 	public:
 		int getCoeff();
 		void setCoeff(int _coeff);
-		virtual int eval(Piece* _chord) = 0;
+		virtual float eval(Piece* _chord) = 0;
 };
 
 #endif //ABSTRACT_PIECE_CONSTRAINT_H
