+++ src/models/Piece.h
@@ -21,7 +21,7 @@ class Piece
 		float tempo;
 		int numberOfChords;
 		Chord* chords[CHORD];
-		int mark;
+		float mark;
 		static vector<AbstractPieceConstraint*> constraints;
 
 	public:
@@ -36,8 +36,8 @@ class Piece
 		void setNumberOfChords(int _numberOfChords);
 		Chord* getChord(int _i);
 		void setChord(int _i, Chord* _c);
-		int getMark();
-		void setMark(int _mark);
+		float getMark();
+		void setMark(float _mark);
 		Piece* mutate(Piece* _piece);
 		template <class E>
 		static AbstractPieceConstraint* getConstraint();
